requirements.txt
matplotlib
python-dotenv
requests
seaborn
SQLAlchemy